tests creating posts to other testcase
